+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/navigation-left/index.tsx
