+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/navigation-left/navigation-left.tsx
@@ -14,11 +14,11 @@
  **/
 import * as React from 'react'
 import styled from 'styled-components'
-import UnsavedIndicator from '../unsaved-indicator'
-const SlideoutLeftViewInstance = require('../../../component/singletons/slideout.left.view-instance.js')
-import ExtensionPoints from '../../../extension-points'
-import { Button, buttonTypeEnum } from '../button'
-const HandlebarsHelpers = require('../../../js/HandlebarsHelpers')
+import UnsavedIndicator from '../presentation/unsaved-indicator'
+const SlideoutLeftViewInstance = require('../../component/singletons/slideout.left.view-instance.js')
+import ExtensionPoints from '../../extension-points'
+import { Button, buttonTypeEnum } from '../presentation/button'
+const HandlebarsHelpers = require('../../js/HandlebarsHelpers')
 import { hot } from 'react-hot-loader'
 
 const { createAction } = require('imperio')
