+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/navigation/presentation.tsx
@@ -17,7 +17,7 @@ import styled from 'styled-components'
 import { CustomElement } from '../styles/mixins'
 import { ChangeBackground } from '../styles/mixins'
 import { transparentize, readableColor } from 'polished'
-import NavigationLeftComponent from '../presentation/navigation-left'
+import NavigationLeftComponent from '../navigation-left'
 import NavigationRightComponent from '../container/navigation-right-container'
 import CancelDrawingContainer from '../container/cancel-drawing-container'
 import Routes from '../container/routes-container'
